+++ tests/testthat/test-reactable.R
@@ -132,6 +132,14 @@ test_that("data with numbers are serialized with max precision", {
   expect_equal(as.character(attribs$data), '{"x":[0.123456789012345]}')
 })
 
+test_that("data with numeric NA, NaN, Inf, and -Inf are serialized correctly", {
+  # jsonlite::toJSON(na = "null") will serialize NA as null, but also discard NaN, Inf, and -Inf
+  data <- data.frame(n = c(1, NA, NaN, Inf, -Inf, 0, -1))
+  tbl <- reactable(data)
+  attribs <- getAttribs(tbl)
+  expect_equal(as.character(attribs$data), '{"n":[1,"NA","NaN","Inf","-Inf",0,-1]}')
+})
+
 test_that("data with dates/datetimes are serialized in ISO 8601", {
   data <- data.frame(x = as.POSIXct("2019-05-06 3:22:15", tz = "UTC"), y = as.Date("2010-12-30"))
   tbl <- reactable(data)